refacto code to fix toolbar behavior
