--- openaev-front/src/admin/components/common/toolBar/BulkToolBar-model.ts
@@ -1,7 +1,0 @@
-import type { ReactElement } from 'react';
-
-export interface ToolTasks {
-  type: string;
-  icon: () => ReactElement;
-  function: () => void;
-}
